requirements.txt
# Django
Django>=5.0,<6.0
gunicorn>=21.2.0

# Database
psycopg[binary]>=3.1.0

# Cache & Celery
redis>=5.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Authentication
django-allauth[socialaccount]>=0.60.0
PyJWT>=2.8.0
cryptography>=41.0.0
django-two-factor-auth[phonenumbers]>=1.16.0
qrcode>=7.4.0

# Static files
whitenoise>=6.6.0

# Markdown & content
Markdown>=3.5.0
Pygments>=2.17.0
django-ckeditor-5>=0.2.10
django-modeltranslation>=0.18.0

# HTTP requests (for API/scraping tasks)
requests>=2.31.0

# Image processing
Pillow>=10.0.0
